+++ project/src/main/java/com/company/banking/Application.java
@@ -13,7 +13,7 @@ import com.company.banking.strategy.SavingsAccountStrategy;
 public class Application {
 	public static void main(String[] args) {
 
-		BankingService bankingService = new BankingServiceImpl();
+		BankingService bankingService = BankingServiceImpl.getInstance();
 		// create 2 accounts;
 		Customer customer = new Customer("James");
