--- project/src/main/java/com/company/ui/bank/BankingServiceImpSingleton.java
@@ -1,17 +1,0 @@
-package com.company.ui.bank;
-
-import com.company.banking.service.BankingServiceImpl;
-
-public class BankingServiceImpSingleton {
-    private static BankingServiceImpl instance = null;
-
-    private BankingServiceImpSingleton() {
-    }
-
-    public static BankingServiceImpl getInstance() {
-        if (instance == null) {
-            instance = new BankingServiceImpl();
-        }
-        return instance;
-    }
-}
