+++ project/src/main/java/com/company/ui/bank/JDialog_AddCompAcc.java
@@ -23,7 +23,7 @@ public class JDialog_AddCompAcc extends JDialog
 	{
 		super(parent);
 		parentframe=parent;
-		bankingService = BankingServiceImpSingleton.getInstance();
+		bankingService = BankingServiceImpl.getInstance();
 		
 		// This code is automatically generated by Visual Cafe when you add
 		// components to the visual environment. It instantiates and initializes
@@ -31,7 +31,7 @@ public class JDialog_AddCompAcc extends JDialog
 		// what Visual Cafe can generate, or Visual Cafe may be unable to back
 		// parse your Java file into its visual environment.
 		//{{INIT_CONTROLS
-		setTitle("Add compamy account");
+		setTitle("Add company account");
 		setModal(true);
 		getContentPane().setLayout(null);
 		setSize(298,339);
