+++ project/src/main/java/com/company/ui/bank/JDialog_Deposit.java
@@ -18,7 +18,7 @@ public class JDialog_Deposit extends JDialog
 		super(parent);
 		parentframe=parent;
 		accnr=aaccnr;
-		bankingService = BankingServiceImpSingleton.getInstance();
+		bankingService = BankingServiceImpl.getInstance();
 		// This code is automatically generated by Visual Cafe when you add
 		// components to the visual environment. It instantiates and initializes
 		// the components. To modify the code, only use code syntax that matches
