+++ project/src/main/java/com/company/ui/bank/JDialog_Withdraw.java
@@ -19,7 +19,7 @@ public class JDialog_Withdraw extends JDialog
 		super(parent);
 		parentframe=parent;
 		accnr=aaccnr;
-		bankingService = BankingServiceImpSingleton.getInstance();
+		bankingService = BankingServiceImpl.getInstance();
 		
 		// This code is automatically generated by Visual Cafe when you add
 		// components to the visual environment. It instantiates and initializes
